Significantly improved the speed of the syntax high-lighter on long lines.
This was done by creating a new ScalarRule for greedily matching scalars not inside double and single quotes instead of relying on the default rule for this. This reduce the number of times that evaluate() will be called for each rule since we match more characters at the time.

The use of RegexRule was also replaced by more specific rules for each of the tokens that used the RegexRule. The more specific rules should be more efficient since the number of regex matches done on the text has decreased.

git-svn-id: https://yedit.googlecode.com/svn/trunk@106 ef3a9a2a-22d2-11de-96a1-5542d1fc8aa6